@@ -1,6 +1,6 @@
 module github.com/GetStream/stream-chat-go/v3
 
-go 1.14
+go 1.16
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
